refactor(ui): extract MistHelperTUI to src/ui/tui.py
Move the MistHelperTUI class (~2100 lines) out of MistHelper.py into
src/ui/tui.py as part of ongoing modularization (Issue #204).

Changes:
- Add src/ui/tui.py with fully self-contained MistHelperTUI class
- Add src/ui/__init__.py package marker
- Remove MistHelperTUI class from MistHelper.py
- Replace module-level import with lazy imports at usage sites
- Remove now-unused 'import platform' from MistHelper.py
- Add src/ui/tui.py to radon exclusion list (high CC, legacy code)
- Add src/ui/tui.py to coverage omit list

Closes #204
